pyproject.toml
[project]
name = "repl-nix-amongusroster"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.36",
    "twilio>=9.4.1",
]
